update: добавил для календаря значения с плавающей точкой

+++ backend/src/main/kotlin/simple_it/models/calendar/calendarDAO/calendarDAO.kt
@@ -10,6 +10,7 @@ import org.jetbrains.exposed.sql.transactions.transaction
 import simple_it.models.calendar.calendarDTO.CreateVacancyCalendar
 import simple_it.models.calendar.calendarDTO.VacancyCalendar
 import simple_it.models.calendar.calendarDTO.VacancyCalendarDTO
+import java.math.BigDecimal
 import java.util.*
 
 class CalendarService(private val database: Database) {
@@ -24,13 +25,16 @@ class CalendarService(private val database: Database) {
         newSuspendedTransaction(Dispatchers.IO) { block() }
 
     suspend fun create(calendar: CreateVacancyCalendar): VacancyCalendarDTO = dbQuery {
-        val slotsMath = Math.floorDiv(calendar.dayEnd - calendar.dayStart, calendar.duration + calendar.freeTime)
+        val slotsMath = Math.floorDiv(
+            (calendar.dayEnd - calendar.dayStart).toInt(),
+            (calendar.duration + calendar.freeTime).toInt()
+        )
 
         val insertStatement = VacancyCalendar.insert {
-            it[duration] = calendar.duration
-            it[freeTime] = calendar.freeTime
-            it[dayStart] = calendar.dayStart
-            it[dayEnd] = calendar.dayEnd
+            it[duration] = BigDecimal.valueOf(calendar.duration)
+            it[freeTime] = BigDecimal.valueOf(calendar.freeTime)
+            it[dayStart] = BigDecimal.valueOf(calendar.dayStart)
+            it[dayEnd] = BigDecimal.valueOf(calendar.dayEnd)
 //            it[freeSlots] = toFreeSlotsCalendarString(calendar.freeSlots)
             it[slots] = slotsMath
             it[maxReservDays] = calendar.maxReservDays
@@ -40,10 +44,10 @@ class CalendarService(private val database: Database) {
         val result = insertStatement.resultedValues?.firstOrNull()
         VacancyCalendarDTO(
             id = result?.get(VacancyCalendar.id) ?: throw Exception("Failed to retrieve inserted calendar ID"),
-            duration = result.get(VacancyCalendar.duration),
-            freeTime = result.get(VacancyCalendar.freeTime),
-            dayStart = result.get(VacancyCalendar.dayStart),
-            dayEnd = result.get(VacancyCalendar.dayEnd),
+            duration = result.get(VacancyCalendar.duration).toDouble(),
+            freeTime = result.get(VacancyCalendar.freeTime).toDouble(),
+            dayStart = result.get(VacancyCalendar.dayStart).toDouble(),
+            dayEnd = result.get(VacancyCalendar.dayEnd).toDouble(),
             slots = result.get(VacancyCalendar.slots),
 //            freeSlots = fromFreeSlotsCalendarString(result.get(VacancyCalendar.freeSlots)) ?: emptyList(),
             maxReservDays = result.get(VacancyCalendar.maxReservDays),
@@ -57,10 +61,10 @@ class CalendarService(private val database: Database) {
             .mapNotNull {
                 VacancyCalendarDTO(
                     id = it[VacancyCalendar.id],
-                    duration = it[VacancyCalendar.duration],
-                    freeTime = it[VacancyCalendar.freeTime],
-                    dayStart = it[VacancyCalendar.dayStart],
-                    dayEnd = it[VacancyCalendar.dayEnd],
+                    duration = it[VacancyCalendar.duration].toDouble(),
+                    freeTime = it[VacancyCalendar.freeTime].toDouble(),
+                    dayStart = it[VacancyCalendar.dayStart].toDouble(),
+                    dayEnd = it[VacancyCalendar.dayEnd].toDouble(),
                     slots = it[VacancyCalendar.slots],
 //                    freeSlots = fromFreeSlotsCalendarString(it[VacancyCalendar.freeSlots]) ?: emptyList(),
                     maxReservDays = it[VacancyCalendar.maxReservDays],
@@ -74,10 +78,10 @@ class CalendarService(private val database: Database) {
         VacancyCalendar.selectAll().map {
             VacancyCalendarDTO(
                 id = it[VacancyCalendar.id],
-                duration = it[VacancyCalendar.duration],
-                freeTime = it[VacancyCalendar.freeTime],
-                dayStart = it[VacancyCalendar.dayStart],
-                dayEnd = it[VacancyCalendar.dayEnd],
+                duration = it[VacancyCalendar.duration].toDouble(),
+                freeTime = it[VacancyCalendar.freeTime].toDouble(),
+                dayStart = it[VacancyCalendar.dayStart].toDouble(),
+                dayEnd = it[VacancyCalendar.dayEnd].toDouble(),
                 slots = it[VacancyCalendar.slots],
 //                freeSlots = fromFreeSlotsCalendarString(it[VacancyCalendar.freeSlots]) ?: emptyList(),
                 maxReservDays = it[VacancyCalendar.maxReservDays],
@@ -89,10 +93,10 @@ class CalendarService(private val database: Database) {
 
     suspend fun update(id: UUID, calendar: CreateVacancyCalendar): Boolean = dbQuery {
         val updatedRows = VacancyCalendar.update({ VacancyCalendar.id eq id }) {
-            it[duration] = calendar.duration
-            it[freeTime] = calendar.freeTime
-            it[dayStart] = calendar.dayStart
-            it[dayEnd] = calendar.dayEnd
+            it[duration] = BigDecimal.valueOf(calendar.duration)
+            it[freeTime] = BigDecimal.valueOf(calendar.freeTime)
+            it[dayStart] = BigDecimal.valueOf(calendar.dayStart)
+            it[dayEnd] = BigDecimal.valueOf(calendar.dayEnd)
 //            it[freeSlots] = toFreeSlotsCalendarString(calendar.freeSlots)
             it[maxReservDays] = calendar.maxReservDays
             it[workingDays] = Json.encodeToString(calendar.workingDays)
